전체 수정 완료, 하지만 삭제 버튼 클릭시 transaction 아이템 1개도 클릭한 걸로 인식해 수정 필요
