+++ src/base-ui/listItem/AmountText/index.jsx
@@ -1,15 +1,22 @@
 import styled from "styled-components";
 
-export default function AmountText({ amount, currentType }) {
+export default function AmountText({
+  earning = null,
+  expense = null,
+  ishover,
+}) {
   return (
-    <Text $currentType={currentType === "earning" ? "income" : "expense"}>
-      {isExpense ? "-" : ""}
-      {Number(amount).toLocaleString()}
+    <Text $ishover={ishover} $currentType={earning ? "income" : "expense"}>
+      {expense ? "-" : ""}
+      {earning ? earning : expense}
     </Text>
   );
 }
 
 const Text = styled.div`
+  transform: ${({ $ishover }) =>
+    $ishover ? `translateX(-57px)` : `translateX(0px)`};
+  transition: transform 0.3s ease;
   width: 186px;
   font-family: ${({ theme }) => theme.font.sans};
   font-weight: 300;
