+++ src/units/DailyTransactionList/index.jsx
@@ -15,7 +15,6 @@ const Header = styled.div`
   justify-content: space-between;
 `;
 const Body = styled.div`
-  position: relative;
   display: flex;
   flex-direction: column;
   border-top: 1px solid ${({ theme }) => theme.color.token.border.default};
